blog/serializers.py
from blog.models import Post

from rest_framework import serializers


class PostSerial(serializers.ModelSerializer):
    owner = serializers.HiddenField(default=serializers.CurrentUserDefault())

    class Meta:
        model = Post
        fields = "__all__"
