requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
django-extensions==3.2.1
django_debug_toolbar==3.8.1
executing==1.2.0
Faker==17.0.0
flake8==5.0.4
flake8-builtins==2.1.0
flake8-django==1.1.5
flake8-import-order==0.18.2
flake8-print==5.0.0
ipython==8.10.0
jedi==0.18.2
kombu==5.2.4
mccabe==0.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.37
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
vine==5.0.0
wcwidth==0.2.6

Django~=4.2.3
djangorestframework~=3.14.0
python-dotenv~=1.0.0